[refactor]: split queue operations app service and preserve review summary wording
