+++ backend/aspnet-core/src/MedStream.Application/Services/QueueOperationsService/QueueOperationsAppService.cs
@@ -1,3 +1,4 @@
+#nullable enable
 using Abp.Authorization;
 using Abp.Authorization.Users;
 using Abp.Domain.Repositories;
@@ -7,26 +8,18 @@ using MedStream.Authorization.Users;
 using MedStream.PatientIntake;
 using MedStream.QueueOperations.Dto;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
-using Newtonsoft.Json;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Text;
-using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace MedStream.QueueOperations;
 
 /// <summary>
-/// Application service for clinician queue listing, review, and status transitions.
+/// Application service for clinician queue listing and review orchestration.
 /// </summary>
 [AbpAuthorize]
-public class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperationsAppService
+public partial class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperationsAppService
 {
     private static readonly IReadOnlyDictionary<string, HashSet<string>> AllowedTransitions =
         new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase)
@@ -60,7 +53,7 @@ public class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperatio
     private readonly IRepository<EncounterNote, long> _encounterNoteRepository;
     private readonly IRepository<User, long> _userRepository;
     private readonly UserManager _userManager;
-    private readonly IConfiguration _configuration;
+    private readonly Microsoft.Extensions.Configuration.IConfiguration _configuration;
     private readonly IQueueRealtimeNotifier _queueRealtimeNotifier;
 
     public QueueOperationsAppService(
@@ -73,7 +66,7 @@ public class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperatio
         IRepository<User, long> userRepository,
         UserManager userManager,
         IQueueRealtimeNotifier queueRealtimeNotifier,
-        IConfiguration configuration = null)
+        Microsoft.Extensions.Configuration.IConfiguration configuration = null)
     {
         _queueTicketRepository = queueTicketRepository;
         _queueEventRepository = queueEventRepository;
@@ -202,18 +195,10 @@ public class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperatio
                 AverageWaitingMinutes = waitingRows.Count == 0
                     ? 0
                     : (int)Math.Round(waitingRows.Average(item => Math.Max(0, (nowUtc - item.EnteredQueueAt).TotalMinutes))),
-                UrgentCount = summaryRows.Count(item =>
-                    item.IsActive &&
-                    string.Equals(item.UrgencyLevel, "Urgent", StringComparison.OrdinalIgnoreCase)),
-                SeenTodayCount = summaryRows.Count(item =>
-                    item.CompletedAt.HasValue &&
-                    item.CompletedAt.Value >= startOfTodayUtc),
-                CalledCount = summaryRows.Count(item =>
-                    item.IsActive &&
-                    string.Equals(item.QueueStatus, PatientIntakeConstants.QueueStatusCalled, StringComparison.OrdinalIgnoreCase)),
-                InConsultationCount = summaryRows.Count(item =>
-                    item.IsActive &&
-                    string.Equals(item.QueueStatus, PatientIntakeConstants.QueueStatusInConsultation, StringComparison.OrdinalIgnoreCase))
+                UrgentCount = summaryRows.Count(item => item.IsActive && string.Equals(item.UrgencyLevel, "Urgent", StringComparison.OrdinalIgnoreCase)),
+                SeenTodayCount = summaryRows.Count(item => item.CompletedAt.HasValue && item.CompletedAt.Value >= startOfTodayUtc),
+                CalledCount = summaryRows.Count(item => item.IsActive && string.Equals(item.QueueStatus, PatientIntakeConstants.QueueStatusCalled, StringComparison.OrdinalIgnoreCase)),
+                InConsultationCount = summaryRows.Count(item => item.IsActive && string.Equals(item.QueueStatus, PatientIntakeConstants.QueueStatusInConsultation, StringComparison.OrdinalIgnoreCase))
             }
         };
     }
@@ -267,8 +252,16 @@ public class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperatio
         var followUpAnswers = DeserializeAnswerDictionary(queueProjection.FollowUpAnswersJson);
         var chiefComplaint = BuildChiefComplaint(queueProjection.ChiefComplaint, selectedSymptoms, extractedPrimarySymptoms, followUpAnswers);
         var reasoning = BuildReasoningItems(queueProjection.Explanation, queueProjection.RedFlagsDetected, followUpAnswers);
-        var clinicianSummary = await BuildClinicianSummaryAsync(chiefComplaint, selectedSymptoms, extractedPrimarySymptoms, followUpAnswers, queueProjection.UrgencyLevel, queueProjection.Explanation, reasoning);
+        var clinicianSummary = await BuildClinicianSummaryAsync(
+            chiefComplaint,
+            selectedSymptoms,
+            extractedPrimarySymptoms,
+            followUpAnswers,
+            queueProjection.UrgencyLevel,
+            queueProjection.Explanation,
+            reasoning);
         var nowUtc = DateTime.UtcNow;
+
         return new ClinicianQueueReviewDto
         {
             QueueTicketId = queueProjection.Id,
@@ -293,701 +286,5 @@ public class QueueOperationsAppService : MedStreamAppServiceBase, IQueueOperatio
             ConsultationPath = $"/clinician/consultation?visitId={queueProjection.VisitId}&queueTicketId={queueProjection.Id}",
             PatientHistoryPath = $"/clinician/history?patientUserId={queueProjection.PatientUserId}&visitId={queueProjection.VisitId}",
         };
-    }
-
-    /// <inheritdoc />
-    public async Task<UpdateQueueTicketStatusOutput> UpdateQueueTicketStatus(UpdateQueueTicketStatusInput input)
-    {
-        var clinician = await EnsureCurrentClinicianAsync();
-        var tenantId = AbpSession.TenantId ?? 1;
-        var facilityId = clinician.ClinicianFacilityId ?? throw new UserFriendlyException("Clinician must be assigned to a facility before updating queue status.");
-
-        var queueTicket = await _queueTicketRepository.FirstOrDefaultAsync(item =>
-            item.TenantId == tenantId &&
-            item.FacilityId == facilityId &&
-            item.Id == input.QueueTicketId &&
-            !item.IsDeleted);
-        if (queueTicket == null)
-        {
-            throw new UserFriendlyException("Queue ticket was not found in your facility context.");
-        }
-
-        var oldStatus = queueTicket.QueueStatus;
-        var normalizedNewStatus = (input.NewStatus ?? string.Empty).Trim().ToLowerInvariant();
-        if (!AllowedTransitions.TryGetValue(oldStatus, out var allowedNewStatuses) || !allowedNewStatuses.Contains(normalizedNewStatus))
-        {
-            throw new UserFriendlyException($"Invalid queue transition from '{oldStatus}' to '{normalizedNewStatus}'.");
-        }
-
-        var changedAt = DateTime.UtcNow;
-        queueTicket.QueueStatus = normalizedNewStatus;
-        queueTicket.CurrentStage = GetStageLabelForStatus(normalizedNewStatus);
-        queueTicket.LastStatusChangedAt = changedAt;
-        queueTicket.CurrentClinicianUserId = clinician.Id;
-
-        if (string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusCalled, StringComparison.OrdinalIgnoreCase))
-        {
-            queueTicket.CalledAt ??= changedAt;
-        }
-
-        if (string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusInConsultation, StringComparison.OrdinalIgnoreCase))
-        {
-            queueTicket.ConsultationStartedAt ??= changedAt;
-            queueTicket.ConsultationStartedByClinicianUserId ??= clinician.Id;
-        }
-
-        if (string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusCompleted, StringComparison.OrdinalIgnoreCase))
-        {
-            queueTicket.CompletedAt = changedAt;
-            queueTicket.IsActive = false;
-        }
-
-        if (string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusCancelled, StringComparison.OrdinalIgnoreCase))
-        {
-            queueTicket.CancelledAt = changedAt;
-            queueTicket.IsActive = false;
-        }
-
-        await _queueTicketRepository.UpdateAsync(queueTicket);
-
-        await _queueEventRepository.InsertAsync(new QueueEvent
-        {
-            TenantId = queueTicket.TenantId,
-            QueueTicketId = queueTicket.Id,
-            EventType = string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusInConsultation, StringComparison.OrdinalIgnoreCase)
-                ? PatientIntakeConstants.QueueEventConsultationStarted
-                : PatientIntakeConstants.QueueEventStatusChanged,
-            OldStatus = oldStatus,
-            NewStatus = normalizedNewStatus,
-            ChangedByClinicianUserId = clinician.Id,
-            Notes = string.IsNullOrWhiteSpace(input.Note)
-                ? $"Queue status changed from '{oldStatus}' to '{normalizedNewStatus}'."
-                : input.Note.Trim(),
-            EventAt = changedAt,
-        });
-
-        var visit = await _visitRepository.GetAsync(queueTicket.VisitId);
-        if (string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusCompleted, StringComparison.OrdinalIgnoreCase))
-        {
-            var encounterNote = await _encounterNoteRepository.FirstOrDefaultAsync(item =>
-                item.TenantId == tenantId &&
-                item.VisitId == visit.Id &&
-                !item.IsDeleted);
-            if (encounterNote == null || !string.Equals(encounterNote.Status, PatientIntakeConstants.EncounterNoteStatusFinalized, StringComparison.OrdinalIgnoreCase))
-            {
-                throw new UserFriendlyException("Encounter note must be finalized before completing the visit.");
-            }
-        }
-
-        if (string.Equals(normalizedNewStatus, PatientIntakeConstants.QueueStatusInConsultation, StringComparison.OrdinalIgnoreCase))
-        {
-            visit.AssignedClinicianUserId ??= clinician.Id;
-        }
-
-        visit.Status = GetVisitStatusForQueueStatus(normalizedNewStatus);
-        await _visitRepository.UpdateAsync(visit);
-
-        await CurrentUnitOfWork.SaveChangesAsync();
-        if (_queueRealtimeNotifier != null)
-        {
-            await _queueRealtimeNotifier.NotifyFacilityQueueChangedAsync(queueTicket.FacilityId);
-            await _queueRealtimeNotifier.NotifyPatientQueueChangedAsync(visit.PatientUserId, visit.Id, queueTicket.Id);
-        }
-
-        return new UpdateQueueTicketStatusOutput
-        {
-            QueueTicketId = queueTicket.Id,
-            OldStatus = oldStatus,
-            NewStatus = queueTicket.QueueStatus,
-            CurrentStage = queueTicket.CurrentStage,
-            ChangedAt = changedAt,
-            VisitId = queueTicket.VisitId,
-            PatientUserId = visit.PatientUserId,
-        };
-    }
-
-    /// <inheritdoc />
-    [HttpPost]
-    public async Task<OverrideQueueTicketUrgencyOutput> OverrideQueueTicketUrgency(OverrideQueueTicketUrgencyInput input)
-    {
-        var clinician = await EnsureCurrentClinicianAsync();
-        var tenantId = AbpSession.TenantId ?? 1;
-        var facilityId = clinician.ClinicianFacilityId ?? throw new UserFriendlyException("Clinician must be assigned to a facility before overriding urgency.");
-        var normalizedUrgencyLevel = NormalizeUrgencyLevel(input.UrgencyLevel);
-
-        var queueTicket = await _queueTicketRepository.FirstOrDefaultAsync(item =>
-            item.TenantId == tenantId &&
-            item.FacilityId == facilityId &&
-            item.Id == input.QueueTicketId &&
-            !item.IsDeleted);
-        if (queueTicket == null)
-        {
-            throw new UserFriendlyException("Queue ticket was not found in your facility context.");
-        }
-
-        var triageAssessment = await _triageAssessmentRepository.GetAsync(queueTicket.TriageAssessmentId);
-        var previousUrgencyLevel = triageAssessment.UrgencyLevel;
-
-        triageAssessment.UrgencyLevel = normalizedUrgencyLevel;
-        triageAssessment.PriorityScore = GetPriorityScoreForUrgency(normalizedUrgencyLevel, triageAssessment.PriorityScore);
-        triageAssessment.Explanation = string.IsNullOrWhiteSpace(input.Note)
-            ? $"Urgency overridden by clinician to {normalizedUrgencyLevel}."
-            : $"{triageAssessment.Explanation} Override: {input.Note.Trim()}".Trim();
-
-        await _triageAssessmentRepository.UpdateAsync(triageAssessment);
-
-        await _queueEventRepository.InsertAsync(new QueueEvent
-        {
-            TenantId = queueTicket.TenantId,
-            QueueTicketId = queueTicket.Id,
-            EventType = PatientIntakeConstants.QueueEventStatusChanged,
-            OldStatus = queueTicket.QueueStatus,
-            NewStatus = queueTicket.QueueStatus,
-            ChangedByClinicianUserId = clinician.Id,
-            Notes = string.IsNullOrWhiteSpace(input.Note)
-                ? $"Urgency overridden from '{previousUrgencyLevel}' to '{normalizedUrgencyLevel}'."
-                : $"Urgency overridden from '{previousUrgencyLevel}' to '{normalizedUrgencyLevel}'. {input.Note.Trim()}",
-            EventAt = DateTime.UtcNow,
-        });
-
-        await CurrentUnitOfWork.SaveChangesAsync();
-        if (_queueRealtimeNotifier != null)
-        {
-            await _queueRealtimeNotifier.NotifyFacilityQueueChangedAsync(queueTicket.FacilityId);
-            await _queueRealtimeNotifier.NotifyPatientQueueChangedAsync((await _visitRepository.GetAsync(queueTicket.VisitId)).PatientUserId, queueTicket.VisitId, queueTicket.Id);
-        }
-
-        return new OverrideQueueTicketUrgencyOutput
-        {
-            QueueTicketId = queueTicket.Id,
-            VisitId = queueTicket.VisitId,
-            PatientUserId = (await _visitRepository.GetAsync(queueTicket.VisitId)).PatientUserId,
-            PreviousUrgencyLevel = previousUrgencyLevel,
-            UrgencyLevel = triageAssessment.UrgencyLevel,
-            PriorityScore = triageAssessment.PriorityScore,
-        };
-    }
-
-    private async Task<User> EnsureCurrentClinicianAsync()
-    {
-        if (!AbpSession.UserId.HasValue)
-        {
-            throw new AbpAuthorizationException("Unauthenticated.");
-        }
-
-        var user = await _userRepository.GetAsync(AbpSession.UserId.Value);
-        var roleNames = await _userManager.GetRolesAsync(user);
-        if (!roleNames.Contains(StaticRoleNames.Tenants.Clinician))
-        {
-            throw new AbpAuthorizationException("Only clinicians may access queue operations.");
-        }
-
-        if (user.IsClinicianApprovalPending)
-        {
-            throw new AbpAuthorizationException("Clinician approval is pending.");
-        }
-
-        return user;
-    }
-
-    private static string GetStageLabelForStatus(string queueStatus)
-    {
-        return queueStatus.ToLowerInvariant() switch
-        {
-            PatientIntakeConstants.QueueStatusWaiting => "Waiting",
-            PatientIntakeConstants.QueueStatusCalled => "Called",
-            PatientIntakeConstants.QueueStatusInConsultation => "In Consultation",
-            PatientIntakeConstants.QueueStatusCompleted => "Completed",
-            PatientIntakeConstants.QueueStatusCancelled => "Cancelled",
-            _ => "Waiting"
-        };
-    }
-
-    private static string GetVisitStatusForQueueStatus(string queueStatus)
-    {
-        return queueStatus.ToLowerInvariant() switch
-        {
-            PatientIntakeConstants.QueueStatusInConsultation => "InConsultation",
-            PatientIntakeConstants.QueueStatusCompleted => "Completed",
-            PatientIntakeConstants.QueueStatusCancelled => "Cancelled",
-            _ => PatientIntakeConstants.VisitStatusQueued
-        };
-    }
-
-    private static HashSet<string> NormalizeFilterValues(IEnumerable<string> values)
-    {
-        return (values ?? Array.Empty<string>())
-            .Where(item => !string.IsNullOrWhiteSpace(item))
-            .Select(item => item.Trim().ToLowerInvariant())
-            .ToHashSet(StringComparer.Ordinal);
-    }
-
-    private static string NormalizeUrgencyLevel(string urgencyLevel)
-    {
-        return urgencyLevel?.Trim().ToLowerInvariant() switch
-        {
-            "urgent" => "Urgent",
-            "priority" => "Priority",
-            "routine" => "Routine",
-            _ => throw new UserFriendlyException("Urgency override must be Urgent, Priority, or Routine.")
-        };
-    }
-
-    private static decimal GetPriorityScoreForUrgency(string urgencyLevel, decimal existingPriorityScore)
-    {
-        return urgencyLevel switch
-        {
-            "Urgent" => Math.Max(existingPriorityScore, 90m),
-            "Priority" => Math.Clamp(existingPriorityScore, 60m, 89m),
-            "Routine" => Math.Min(existingPriorityScore, 59m),
-            _ => existingPriorityScore
-        };
-    }
-
-    private async Task<string> BuildClinicianSummaryAsync(
-        string chiefComplaint,
-        IReadOnlyList<string> selectedSymptoms,
-        IReadOnlyList<string> extractedPrimarySymptoms,
-        IReadOnlyDictionary<string, object> followUpAnswers,
-        string urgencyLevel,
-        string triageExplanation,
-        IReadOnlyList<string> reasoning)
-    {
-        var followUpHighlights = BuildFollowUpHighlights(followUpAnswers);
-        var fallbackSummary = BuildDeterministicClinicianSummary(
-            chiefComplaint,
-            selectedSymptoms,
-            extractedPrimarySymptoms,
-            followUpHighlights,
-            urgencyLevel,
-            triageExplanation,
-            reasoning);
-        var apiKey = GetOpenAiSetting("OpenAI:ApiKey") ?? Environment.GetEnvironmentVariable("OPENAI_API_KEY");
-        if (string.IsNullOrWhiteSpace(apiKey))
-        {
-            return fallbackSummary;
-        }
-
-        var endpoint = (GetOpenAiSetting("OpenAI:BaseUrl") ?? "https://api.openai.com/v1").TrimEnd('/') + "/chat/completions";
-        var model = GetOpenAiSetting("OpenAI:Model") ?? "gpt-4o-mini";
-
-        using var httpClient = new HttpClient();
-        httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", apiKey);
-
-        var payload = new
-        {
-            model,
-            temperature = 0.1,
-            response_format = new { type = "json_object" },
-            messages = new object[]
-            {
-                new
-                {
-                    role = "system",
-                    content = "Write a concise clinician-facing intake summary. Return JSON only with key summary. Mention the main complaint, key symptoms, relevant positives, and any urgent context. Do not diagnose. Keep it to 2-4 short sentences."
-                },
-                new
-                {
-                    role = "user",
-                    content = JsonConvert.SerializeObject(new
-                    {
-                        chiefComplaint,
-                        selectedSymptoms,
-                        extractedPrimarySymptoms,
-                        followUpHighlights,
-                        urgencyLevel,
-                        triageExplanation,
-                        reasoning
-                    })
-                }
-            }
-        };
-
-        try
-        {
-            using var content = new StringContent(JsonConvert.SerializeObject(payload), Encoding.UTF8, "application/json");
-            using var response = await httpClient.PostAsync(endpoint, content);
-            if (!response.IsSuccessStatusCode)
-            {
-                return fallbackSummary;
-            }
-
-            var responseJson = await response.Content.ReadAsStringAsync();
-            using var document = JsonDocument.Parse(responseJson);
-            var completionContent = document.RootElement
-                .GetProperty("choices")[0]
-                .GetProperty("message")
-                .GetProperty("content")
-                .GetString();
-            if (string.IsNullOrWhiteSpace(completionContent))
-            {
-                return fallbackSummary;
-            }
-
-            using var summaryDocument = JsonDocument.Parse(completionContent);
-            if (!summaryDocument.RootElement.TryGetProperty("summary", out var summaryElement))
-            {
-                return fallbackSummary;
-            }
-
-            var summary = summaryElement.GetString()?.Trim();
-            return string.IsNullOrWhiteSpace(summary) ? fallbackSummary : summary;
-        }
-        catch
-        {
-            return fallbackSummary;
-        }
-    }
-
-    private static string BuildDeterministicClinicianSummary(
-        string chiefComplaint,
-        IReadOnlyList<string> selectedSymptoms,
-        IReadOnlyList<string> extractedPrimarySymptoms,
-        IReadOnlyList<string> followUpHighlights,
-        string urgencyLevel,
-        string triageExplanation,
-        IReadOnlyList<string> reasoning)
-    {
-        var summaryParts = new List<string>();
-        if (!string.IsNullOrWhiteSpace(chiefComplaint))
-        {
-            summaryParts.Add($"Primary concern: {chiefComplaint.Trim()}.");
-        }
-
-        var symptomSource = selectedSymptoms.Count > 0 ? selectedSymptoms : extractedPrimarySymptoms;
-        if (symptomSource.Count > 0)
-        {
-            summaryParts.Add($"Symptoms reported: {string.Join(", ", symptomSource)}.");
-        }
-
-        if (followUpHighlights.Count > 0)
-        {
-            summaryParts.Add($"Key positives: {string.Join("; ", followUpHighlights.Take(3))}.");
-        }
-
-        if (!string.IsNullOrWhiteSpace(urgencyLevel))
-        {
-            summaryParts.Add($"Triage category: {urgencyLevel}.");
-        }
-
-        if (!string.IsNullOrWhiteSpace(triageExplanation))
-        {
-            summaryParts.Add(triageExplanation.Trim().TrimEnd('.') + ".");
-        }
-        else if (reasoning.Count > 0)
-        {
-            summaryParts.Add($"Reasoning: {string.Join("; ", reasoning.Take(3))}.");
-        }
-
-        return string.Join(" ", summaryParts.Where(item => !string.IsNullOrWhiteSpace(item))).Trim();
-    }
-
-    private static List<string> BuildFollowUpHighlights(IReadOnlyDictionary<string, object> followUpAnswers)
-    {
-        return followUpAnswers
-            .Where(item => IsPositiveAnswer(item.Value))
-            .Select(item => BuildReadableFollowUpHighlight(item.Key, item.Value))
-            .Where(item => !string.IsNullOrWhiteSpace(item))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToList();
-    }
-
-    private static string BuildReadableFollowUpHighlight(string key, object answer)
-    {
-        var label = GetAnswerLabel(key);
-        if (answer is bool booleanValue)
-        {
-            return booleanValue ? label : string.Empty;
-        }
-
-        var formattedValue = FormatAnswerValue(answer);
-        return string.IsNullOrWhiteSpace(formattedValue) ? string.Empty : $"{label}: {formattedValue}";
-    }
-
-    private static List<string> BuildReasoningItems(string triageExplanation, string serializedRedFlags, IReadOnlyDictionary<string, object> followUpAnswers)
-    {
-        var reasoningItems = new List<string>();
-        if (!string.IsNullOrWhiteSpace(triageExplanation))
-        {
-            reasoningItems.Add(triageExplanation.Trim());
-        }
-
-        var redFlags = DeserializeStringList(serializedRedFlags);
-        reasoningItems.AddRange(redFlags
-            .Select(item => MapReasoningCode(item, followUpAnswers))
-            .Where(item => !string.IsNullOrWhiteSpace(item)));
-
-        var urgentCheckReasons = new[]
-        {
-            "urgentSevereBreathing",
-            "urgentSevereChestPain",
-            "urgentUncontrolledBleeding",
-            "urgentCollapse",
-            "urgentConfusion"
-        }
-        .Where(key => GetBooleanAnswer(followUpAnswers, key))
-        .Select(GetAnswerLabel);
-
-        reasoningItems.AddRange(urgentCheckReasons);
-
-        return reasoningItems
-            .Where(item => !string.IsNullOrWhiteSpace(item))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToList();
-    }
-
-    private static string BuildChiefComplaint(
-        string freeTextComplaint,
-        IReadOnlyList<string> selectedSymptoms,
-        IReadOnlyList<string> extractedPrimarySymptoms,
-        IReadOnlyDictionary<string, object> followUpAnswers)
-    {
-        if (!string.IsNullOrWhiteSpace(freeTextComplaint))
-        {
-            return freeTextComplaint.Trim();
-        }
-
-        if (TryGetStringAnswer(followUpAnswers, "mainConcern", out var mainConcern))
-        {
-            return mainConcern;
-        }
-
-        var urgentComplaint = BuildUrgentChiefComplaint(followUpAnswers);
-        if (!string.IsNullOrWhiteSpace(urgentComplaint))
-        {
-            return urgentComplaint;
-        }
-
-        var firstSelectedSymptom = selectedSymptoms?.FirstOrDefault(item => !string.IsNullOrWhiteSpace(item));
-        if (!string.IsNullOrWhiteSpace(firstSelectedSymptom))
-        {
-            return firstSelectedSymptom.Trim();
-        }
-
-        var firstExtractedSymptom = extractedPrimarySymptoms?.FirstOrDefault(item => !string.IsNullOrWhiteSpace(item));
-        if (!string.IsNullOrWhiteSpace(firstExtractedSymptom))
-        {
-            return firstExtractedSymptom.Trim();
-        }
-
-        return string.Empty;
-    }
-
-    private static string BuildUrgentChiefComplaint(IReadOnlyDictionary<string, object> followUpAnswers)
-    {
-        var urgentConcerns = new List<string>();
-        if (GetBooleanAnswer(followUpAnswers, "urgentSevereBreathing"))
-        {
-            urgentConcerns.Add("Severe breathing difficulty");
-        }
-
-        if (GetBooleanAnswer(followUpAnswers, "urgentSevereChestPain"))
-        {
-            urgentConcerns.Add("Severe chest pain");
-        }
-
-        if (GetBooleanAnswer(followUpAnswers, "urgentUncontrolledBleeding"))
-        {
-            urgentConcerns.Add("Uncontrolled bleeding");
-        }
-
-        if (GetBooleanAnswer(followUpAnswers, "urgentCollapse"))
-        {
-            urgentConcerns.Add("Collapse or blackout");
-        }
-
-        if (GetBooleanAnswer(followUpAnswers, "urgentConfusion"))
-        {
-            urgentConcerns.Add("Acute confusion");
-        }
-
-        return urgentConcerns.Count > 0 ? string.Join(", ", urgentConcerns) : string.Empty;
-    }
-
-    private static List<string> DeserializeStringList(string serializedList)
-    {
-        if (string.IsNullOrWhiteSpace(serializedList))
-        {
-            return new List<string>();
-        }
-
-        try
-        {
-            return JsonConvert.DeserializeObject<List<string>>(serializedList) ?? new List<string>();
-        }
-        catch
-        {
-            return new List<string>();
-        }
-    }
-
-    private static Dictionary<string, object> DeserializeAnswerDictionary(string serializedDictionary)
-    {
-        if (string.IsNullOrWhiteSpace(serializedDictionary))
-        {
-            return new Dictionary<string, object>(StringComparer.OrdinalIgnoreCase);
-        }
-
-        try
-        {
-            return JsonConvert.DeserializeObject<Dictionary<string, object>>(serializedDictionary)
-                   ?? new Dictionary<string, object>(StringComparer.OrdinalIgnoreCase);
-        }
-        catch
-        {
-            return new Dictionary<string, object>(StringComparer.OrdinalIgnoreCase);
-        }
-    }
-
-    private static string MapReasoningCode(string code, IReadOnlyDictionary<string, object> followUpAnswers)
-    {
-        return code?.Trim() switch
-        {
-            "global_urgent_check_positive" => BuildUrgentReasoningFromAnswers(followUpAnswers),
-            "urgent_global_red_flag" => BuildUrgentReasoningFromAnswers(followUpAnswers),
-            "urgent_keyword_detected" => "Complaint text included urgent warning signs.",
-            _ => HumanizeReasoningCode(code)
-        };
-    }
-
-    private static string BuildUrgentReasoningFromAnswers(IReadOnlyDictionary<string, object> followUpAnswers)
-    {
-        var labels = new[]
-        {
-            "urgentSevereBreathing",
-            "urgentSevereChestPain",
-            "urgentUncontrolledBleeding",
-            "urgentCollapse",
-            "urgentConfusion"
-        }
-        .Where(key => GetBooleanAnswer(followUpAnswers, key))
-        .Select(GetAnswerLabel)
-        .ToList();
-
-        return labels.Count > 0
-            ? $"Emergency safety checks were positive: {string.Join(", ", labels)}."
-            : "Emergency safety checks were positive.";
-    }
-
-    private static string HumanizeReasoningCode(string code)
-    {
-        if (string.IsNullOrWhiteSpace(code))
-        {
-            return string.Empty;
-        }
-
-        var normalized = code.Replace("_", " ").Replace("-", " ").Trim();
-        return char.ToUpperInvariant(normalized[0]) + normalized.Substring(1);
-    }
-
-    private static bool TryGetStringAnswer(IReadOnlyDictionary<string, object> answers, string key, out string value)
-    {
-        value = string.Empty;
-        if (answers == null || !answers.TryGetValue(key, out var answer) || answer == null)
-        {
-            return false;
-        }
-
-        value = FormatAnswerValue(answer);
-        return !string.IsNullOrWhiteSpace(value);
-    }
-
-    private static bool GetBooleanAnswer(IReadOnlyDictionary<string, object> answers, string key)
-    {
-        if (answers == null || !answers.TryGetValue(key, out var answer) || answer == null)
-        {
-            return false;
-        }
-
-        if (answer is bool booleanValue)
-        {
-            return booleanValue;
-        }
-
-        if (bool.TryParse(Convert.ToString(answer), out var parsedBoolean))
-        {
-            return parsedBoolean;
-        }
-
-        return false;
-    }
-
-    private static bool IsPositiveAnswer(object answer)
-    {
-        if (answer == null)
-        {
-            return false;
-        }
-
-        if (answer is bool booleanValue)
-        {
-            return booleanValue;
-        }
-
-        if (answer is string stringValue)
-        {
-            return !string.IsNullOrWhiteSpace(stringValue) &&
-                   !string.Equals(stringValue, "false", StringComparison.OrdinalIgnoreCase) &&
-                   !string.Equals(stringValue, "[]", StringComparison.OrdinalIgnoreCase);
-        }
-
-        if (answer is IEnumerable enumerable && answer is not string)
-        {
-            return enumerable.Cast<object>().Any();
-        }
-
-        return true;
-    }
-
-    private static string GetAnswerLabel(string key)
-    {
-        return key switch
-        {
-            "urgentSevereBreathing" => "Severe difficulty breathing reported",
-            "urgentSevereChestPain" => "Severe chest pain reported",
-            "urgentUncontrolledBleeding" => "Uncontrolled bleeding reported",
-            "urgentCollapse" => "Collapse or blackout reported",
-            "urgentConfusion" => "Confusion or reduced responsiveness reported",
-            "mainConcern" => "Main concern",
-            "symptomDurationDays" => "Duration",
-            "durationDays" => "Duration",
-            "pregnant" => "Pregnancy reported",
-            "vomiting" => "Vomiting reported",
-            "severePain" => "Severe pain reported",
-            "yellowEyesSkin" => "Jaundice reported",
-            "noStoolWind" => "Unable to pass stool or wind",
-            "fainting" => "Fainting reported",
-            _ => HumanizeReasoningCode(key)
-        };
-    }
-
-    private static string FormatAnswerValue(object answer)
-    {
-        if (answer == null)
-        {
-            return string.Empty;
-        }
-
-        if (answer is bool booleanValue)
-        {
-            return booleanValue ? "Yes" : "No";
-        }
-
-        if (answer is IEnumerable enumerable && answer is not string)
-        {
-            var values = enumerable
-                .Cast<object>()
-                .Select(item => Convert.ToString(item)?.Trim())
-                .Where(item => !string.IsNullOrWhiteSpace(item))
-                .ToList();
-            return values.Count == 0 ? string.Empty : string.Join(", ", values);
-        }
-
-        return Convert.ToString(answer)?.Trim() ?? string.Empty;
-    }
-
-    private string GetOpenAiSetting(string key)
-    {
-        return _configuration?[key];
     }
 }
